--- logging/LogLevel.java
@@ -1,9 +1,0 @@
-package com.airline.logging;
-
-// Enumeration representing different log levels
-public enum LogLevel {
-    INFO,     // Informational messages
-    WARNING,  // Warnings
-    ERROR,    // Errors
-    DEBUG     // Debugging information
-}
